Encounter - moving everything into Encounter.py while I restructure design
